libalarm: make cookie_t int32_t
The assertion in src/server.c was failing on 64 bit systems, where
sizeof(dbus_int32_t) != sizeof(cookie_t)

+++ src/libalarm.h
@@ -163,7 +163,7 @@ extern "C" {
 
 /** \brief Unique identifier type for the alarm events.
  **/
-typedef long cookie_t;
+typedef int32_t cookie_t;
 
 /* ------------------------------------------------------------------------- *
  * alarm_attr_t
